+++ src/components/megalinks/components/megalink-multi-image/MegalinkMultiImage.vue
@@ -222,8 +222,7 @@ export default {
         }
 
         .vs-stretched-link-card__video-button {
-            position: absolute;
-            bottom: 100%;
+            bottom: 0;
             left: 0;
         }
 
@@ -231,22 +230,7 @@ export default {
 
         @include media-breakpoint-up(xl) {
             &.vs-megalink-multi-image-card--featured {
-                .vs-stretched-link-card__video-button {
-                    bottom: 100%;
-                    left: $spacer-0;
-                }
-
                 @include full-rectangle-video-button();
-            }
-        }
-
-        @include media-breakpoint-up(xl) {
-            &.vs-megalink-multi-image-card--featured {
-                .vs-stretched-link-card__video-button {
-                    bottom: $spacer-050;
-                    left: $spacer-050;
-                    z-index: 2;
-                }
             }
         }
     };
@@ -302,10 +286,6 @@ export default {
                 box-shadow: $vs-elevation-shadow-raised;
             }
 
-            .vs-stretched-link-card__video-button {
-                position: fixed;
-            }
-
             .stretched-link:after {
                 position: fixed;
                 left: 0;
@@ -334,11 +314,6 @@ export default {
                 .card-body {
                     left: $spacer-0;;
                     right: auto;
-                }
-
-                .vs-stretched-link-card__video-button {
-                    left: auto;
-                    right: $spacer-050;
                 }
             }
         }