+++ stories/MegalinksLinkList.stories.js
@@ -5,7 +5,7 @@ import VsCaption from '@/components/caption/Caption.vue';
 import VsLink from '@/components/link/Link.vue';
 import VsModal from '@/components/modal/Modal.vue';
 import VsVideo from '@/components/video/Video.vue';
-import VsBadge from '@/components/badge/VsBadge.vue';
+import VsBadge from '@/components/badge/Badge.vue';
 import {
     VsRow,
     VsCol,